Fixed last checkin variable.
In the HTML source, the last_checkin variable is normally a one
item list consisting of the time since last checkin.

If the time since last checkin is longer than about 30 minutes,
it will be a two item list. The first item is the string "Late!" or
"Down!", and the second item is the time since last checkin.

The important information here is the actual time, which is always
the last item in the list.