Changed data collection method.
Prior to this commit, we would just scrape the CloudTrax usage
information for the nodes and users. While this is easy, it was
not as *nice* as it should be.

CloudTrax calculates usage on a rolling 24hr period. We can deal
with that at the moment.

CloudTrax displays user download and upload statistics in KB. That
is also fine.

CloudTrax displays node download and upload statistics in MB. That
gives us a rounding problem when trying to reconcile user and node
usage.

Add to this that CloudTrax has recently also changed node usage to
be "recent" instead of 24hr giving us futher inconsistencies
between users and nodes.

Now, we collect the node information without usage statistics. We
then collect the user information and iterate through them to
allocate user usage to nodes. The only problem with this approach
is that if a user moves from one node to another, part of their
usage might get allocated to the wrong node.