Node names are optional
In CloudTrax, node names are optional, so the CloudScraper code
tries to allow this. I originally wanted to use the node mac address
as the dictionary key for the nodes. This is not possible as the mac
address that is listed against the user is the wifi mac address, and
the mac address that is listed against the node is the ethernet mac
address. Unfortunately this means we must use the network name as a
dictionary key.