Workaround for roaming user issue
There is a bug that appears if the CloudTrax text in the "Last Seen On"
node report is blank. When this bug is triggered, CloudScraper would
not know which gateway to assign the usage and therefore it would crash.

This does not fix the bug, but does stop a the whole process from crashing.