Use node mac address instead of name
CloudTrax reporting has a few interesting artifacts. When a wifi node
is referenced, they use the ethernet mac address to identify it. When
a wifi user is referenced, the node wifi mac address is used. After
some discussion with the CloudTrax developers, it has been confirmed
that

(ethernet mac address) + 7 = (wifi mac addresS)

This commit uses the ethernet mac address for all addressing.