Implement CloudTrax API Authentication.
The new method CloudTrax.request() will be used for all requests made to
the API.  It is responsible for formulating the Authorization and
Signature headers, using the API key and secret from the configuration.

This is very much a work-in-progress commit, and breaks pretty much
everything.  CloudTrax.request() works great, though.