Clean up and simply CLI script.
Remove options that are not currently supported -- we are targeting a
simple scrape-and-store strategy, where we collect info from the API and
then put it in a database.  Reporting is a story for another day.

Make lib.config.Config a subclass of RawConfigParser, this reduces the
amount of code needed to get things done.